+++ EssentialFeed/EssentialFeediOSTests/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,40 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Prabhat Tiwari on 08/07/25.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension FeedImageCell {
+    
+    func simulateRetryAction() {
+        feedImageRetryButton.simulateTap()
+    }
+    
+    var isShowingLocation: Bool {
+        return !locationContainer.isHidden
+    }
+    
+    var isShowingImageLoadingIndicator: Bool {
+        return feedImageContainer.isShimmering
+    }
+    
+    var locationText: String? {
+        return locationLabel.text
+    }
+    
+    var descriptionText: String? {
+        return descriptionLabel.text
+    }
+    
+    var renderedImage: Data? {
+        return feedImageView.image?.pngData()
+    }
+    
+    var isShowingRetryAction: Bool {
+        return !feedImageRetryButton.isHidden
+    }
+}
